requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
slack-bolt==1.18.1
python-dotenv==1.0.0
apscheduler==3.10.4
